+++ src/mruby_integration/models/window.cpp
@@ -129,6 +129,12 @@ mrb_Window_set_exit_key(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_nil_value();
 }
 
+auto
+mrb_Window_resized(mrb_state*, mrb_value) -> mrb_value
+{
+  return mrb_bool_value(IsWindowResized());
+}
+
 void
 append_models_Window(mrb_state* mrb)
 {
@@ -158,6 +164,8 @@ append_models_Window(mrb_state* mrb)
     mrb, Window_class, "clear_flag", mrb_Window_clear_flag, MRB_ARGS_REQ(1));
   mrb_define_class_method(
     mrb, Window_class, "exit_key=", mrb_Window_set_exit_key, MRB_ARGS_REQ(1));
+  mrb_define_class_method(
+    mrb, Window_class, "resized?", mrb_Window_resized, MRB_ARGS_NONE());
 
   load_ruby_models_window(mrb);
 }